Update to the chat module in an attempt to make the strain on the database lighter - removed the usage of the DBPlayerData. (column in table has also been changed from fnUserId to fcPlayername)
git-svn-id: http://www.twcore.org/svn/trunk/twcore@1408 9f8b3893-2a26-0410-9d00-a71e40fd2f77

+++ src/twcore/bots/pubhub/pubhubchat.java
@@ -33,38 +33,32 @@ public class pubhubchat extends PubBotModule {
 	public void handleEvent(InterProcessEvent event) {
 		if(event.getObject() instanceof pubbotchatIPC) {
 			pubbotchatIPC ipc = (pubbotchatIPC)event.getObject();
+			String arena = Tools.addSlashesToString(ipc.getArena());
+			String message = Tools.addSlashesToString(ipc.getMessage());
+			String sender = (ipc.getMessageType() == Message.ARENA_MESSAGE) ? null : ipc.getSender();
 			
-			DBPlayerData playerdata = new DBPlayerData(m_botAction, dbConn, ipc.getSender());
-			
-			if(playerdata.checkPlayerExists() || ipc.getMessageType() == Message.ARENA_MESSAGE) {
-				String arena = Tools.addSlashesToString(ipc.getArena());
-				String message = Tools.addSlashesToString(ipc.getMessage());
-				String sender = (ipc.getMessageType() == Message.ARENA_MESSAGE) ? null : String.valueOf(playerdata.getUserID());
-				
-				// Save the message
-				try {
-					m_botAction.SQLQuery( dbConn, "INSERT INTO tblChat (`fcArena`, `fnMessageType`, `fnUserID`, `fcMessage`) VALUES ('"+arena+"', "+ipc.getMessageType()+", "+sender+", '"+message+"')" );
-				} catch(SQLException sqle) {
-					Tools.printStackTrace(sqle);
-				}
-				
-				// Clean up
-				try {
-					ResultSet resultSet = m_botAction.SQLQuery( dbConn, "SELECT fnChatId FROM tblChat WHERE fcArena = '"+arena+"'");
-					if(resultSet != null) {
-						resultSet.afterLast();					// Go to the last row
-						resultSet.relative(-maxSavedLines);		// Move up a number of rows to save
-						if(resultSet.isBeforeFirst() == false) {
-							int fnChatID = resultSet.getInt("fnChatID"); // Get the fnChatID of this row
-							m_botAction.SQLQuery( dbConn, "DELETE FROM tblChat WHERE fcArena = '"+arena+"' AND fnChatID < "+fnChatID);
-						}
-					}
-				} catch (SQLException sqle) {
-					Tools.printStackTrace(sqle);
-				}
-				
+			// Save the message
+			try {
+				m_botAction.SQLQuery( dbConn, "INSERT INTO tblChat (`fcArena`, `fnMessageType`, `fcPlayername`, `fcMessage`) VALUES ('"+arena+"', "+ipc.getMessageType()+", '"+sender+"', '"+message+"')" );
+			} catch(SQLException sqle) {
+				Tools.printStackTrace(sqle);
 			}
 			
+			// Clean up
+			try {
+				ResultSet resultSet = m_botAction.SQLQuery( dbConn, "SELECT fnChatId FROM tblChat WHERE fcArena = '"+arena+"'");
+				if(resultSet != null && resultSet.next()) {
+					resultSet.afterLast();					// Go to the last row
+					resultSet.relative(-maxSavedLines);		// Move up a number of rows to save
+					if(resultSet.isBeforeFirst() == false && resultSet.isAfterLast() == false) {
+						int fnChatID = resultSet.getInt("fnChatID"); // Get the fnChatID of this row
+						m_botAction.SQLQuery( dbConn, "DELETE FROM tblChat WHERE fcArena = '"+arena+"' AND fnChatID < "+fnChatID);
+					}
+				}
+				resultSet.close();
+			} catch (SQLException sqle) {
+				Tools.printStackTrace(sqle);
+			}
 		}
 	}
